// pages/src/components/ControlPlane.module.css.d.ts
declare const styles: {
  container: string;
  urlInputs: string;
  inputGroup: string;
  label: string;
  required: string;
  optional: string;
  urlInput: string;
  textInput: string;
  actions: string;
  swapButton: string;
  submitButton: string;
  warningBanner: string;
  warningText: string;
  warningDismiss: string;
  slideIn: string;
};

export default styles;